restructured instrument layers getter
