+++ src/layers/olympex-layers/components/instrumentLayers/helpers/flightTrack.js
@@ -1,23 +1,15 @@
-import { flightTrackBaseUrl } from "../../../config";
+import { flightTrackBaseUrl } from "../../../../../config"
 
-function getInstrumentsItem(date, instrumentType) {
-/** 
-* forms a instrument meta for requested instruments in a particular date, i.e. collect the meta of data for various instruments.
-* @param {string} date - YYYY-MM-DD format. Date of the campaign.
-* @param {string} instrumentType - the name of the instrument, whose meta is needed for a particular date.
-* @return {object} A structured instrument meta item.
-*/
-    switch(instrumentType) {
-        case "trackEr2":
-            return flighttrack(date, "er2");
-        case "trackDc8":
-            return flighttrack(date, "dc8");
-        default:
-            return {}
-    }
-}
-
-const flighttrack = (date, aircraft, index) => {
+export default function flighttrack (date, aircraft, index) {
+    /** 
+    * returns structured flight meta item.
+    * @summary If the description is long, write your summary here. Otherwise, feel free to remove this.
+    * @param {date} string - YYYY-MM-DD format. Date of the campaign when the flight took off.
+    * @param {aircraft} string - Type of aircraft used in the field campaign.
+    * @param {index} number (Optional) - Index of the instrument in the list of instruments that are visualization wanted
+    * @return {Object} structured flight meta item.
+    */
+    
     // TODO: how to get start and end time of the campaign for a particular date.????
     if (aircraft == "er2") {
         return {
@@ -47,13 +39,3 @@ const flighttrack = (date, aircraft, index) => {
         }
     }
 }
-
-const dates = ["2017-05-14"];
-const instruments = ["trackEr2", "trackDc8"]
-
-const instrumentLayers = dates.map(date => ({
-        date: date,
-        items: instruments.map((instrum, index) => getInstrumentsItem(date, instrum, index))
-    }));
-
-export default instrumentLayers;
