+++ src/layers/olympex-layers/components/instrumentLayers/helpers/index.js
@@ -1,0 +1,3 @@
+import flighttrack from "./flightTrack";
+
+export { flighttrack }
